feat: create propper gem structure
