+++ bin/rsfano
@@ -1,0 +1,13 @@
+#!/usr/bin/env ruby
+
+require "ruby-shannon-fano/cli"
+
+begin
+  cli = RubyShannonFano::CLI.new
+  cli.run(ARGV)
+rescue => e
+  raise e if $DEBUG
+  STDERR.puts e.message
+  STDERR.puts e.backtrace.join("\n")
+  exit 1
+end
